demo/demo/requirements.txt
gradio>=4.0.0
requests>=2.25.0
pillow>=10.0.0
pandas>=1.3.0
numpy>=1.21.0 